+++ src/main/java/de/uib/configed/ConfigedMain.java
@@ -246,7 +246,7 @@ public class ConfigedMain {
 		ServerActionManager.initData(this);
 		Messagebus.initMessagebus(this);
 
-		clientSearch = new ClientSearch(this);
+		clientSearch = new ClientSearch();
 
 		initialDataLoader = new InitialDataLoader(this);
 		initialDataLoader.execute();